pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faster-whisper>=1.1.1",
    "google-genai>=1.23.0",
    "pytube>=15.0.0",
    "streamlit>=1.46.1",
    "yt-dlp>=2025.6.30",
]
